requirements.txt
albumentations==0.5.2
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imgaug==0.4.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image
scikit-learn
scipy
Shapely==1.7.1
six==1.15.0
threadpoolctl==2.1.0
tifffile
torch
torchvision
tqdm==4.56.0
typing-extensions==3.7.4.3
